+++ dart/lib/src/protocol/sdk_info.dart
@@ -15,6 +15,17 @@ class SdkInfo {
     this.versionPatchlevel,
   });
 
+  /// Deserializes a [SdkInfo] from JSON [Map].
+  factory SdkInfo.fromJson(Map<String, dynamic> json) {
+    return SdkInfo(
+      sdkName: json['sdk_name'],
+      versionMajor: json['version_major'],
+      versionMinor: json['version_minor'],
+      versionPatchlevel: json['version_patchlevel'],
+    );
+  }
+
+  /// Produces a [Map] that can be serialized to JSON.
   Map<String, dynamic> toJson() {
     final json = <String, dynamic>{};
     if (sdkName != null) {