+++ dart/lib/src/protocol/sentry_user.dart
@@ -53,6 +53,17 @@ class SentryUser {
   /// by Sentry.
   final Map<String, dynamic>? extras;
 
+  /// Deserializes a [SentryUser] from JSON [Map].
+  factory SentryUser.fromJson(Map<String, dynamic> json) {
+    return SentryUser(
+      id: json['id'],
+      username: json['username'],
+      email: json['email'],
+      ipAddress: json['ip_address'],
+      extras: json['extras'],
+    );
+  }
+
   /// Produces a [Map] that can be serialized to JSON.
   Map<String, dynamic> toJson() {
     return <String, dynamic>{