+++ dart/lib/src/transport/noop_transport.dart
@@ -7,10 +7,6 @@ import 'transport.dart';
 
 class NoOpTransport implements Transport {
   @override
-  Future<SentryId> sendSentryEvent(SentryEvent event) =>
-      Future.value(SentryId.empty());
-
-  @override
-  Future<SentryId> sendSentryEnvelope(SentryEnvelope envelope) =>
+  Future<SentryId> send(SentryEnvelope envelope) =>
       Future.value(SentryId.empty());
 }