+++ dart/test/transport/http_transport_test.dart
@@ -45,7 +45,7 @@ void main() {
       final sut = fixture.getSut(httpMock, mockRateLimiter);
 
       final sentryEnvelope = givenEnvelope();
-      await sut.sendSentryEnvelope(sentryEnvelope);
+      await sut.send(sentryEnvelope);
 
       expect(mockRateLimiter.envelopeToFilter, sentryEnvelope);
     });
@@ -66,7 +66,9 @@ void main() {
       final sut = fixture.getSut(httpMock, mockRateLimiter);
 
       final sentryEvent = SentryEvent();
-      await sut.sendSentryEvent(sentryEvent);
+      final envelope =
+          SentryEnvelope.fromEvent(sentryEvent, fixture.options.sdk);
+      await sut.send(envelope);
 
       final envelopeData = <int>[];
       await filteredEnvelope.envelopeStream().forEach(envelopeData.addAll);
@@ -86,7 +88,9 @@ void main() {
       final sut = fixture.getSut(httpMock, mockRateLimiter);
 
       final sentryEvent = SentryEvent();
-      final eventId = await sut.sendSentryEvent(sentryEvent);
+      final envelope =
+          SentryEnvelope.fromEvent(sentryEvent, fixture.options.sdk);
+      final eventId = await sut.send(envelope);
 
       expect(eventId, SentryId.empty());
       expect(httpCalled, false);
@@ -108,7 +112,10 @@ void main() {
       final sut = fixture.getSut(httpMock, mockRateLimiter);
 
       final sentryEvent = SentryEvent();
-      await sut.sendSentryEvent(sentryEvent);
+      final envelope =
+          SentryEnvelope.fromEvent(sentryEvent, fixture.options.sdk);
+      await sut.send(envelope);
+
       expect(mockRateLimiter.envelopeToFilter?.header.eventId,
           sentryEvent.eventId);
 
@@ -126,7 +133,9 @@ void main() {
       final sut = fixture.getSut(httpMock, mockRateLimiter);
 
       final sentryEvent = SentryEvent();
-      await sut.sendSentryEvent(sentryEvent);
+      final envelope =
+          SentryEnvelope.fromEvent(sentryEvent, fixture.options.sdk);
+      await sut.send(envelope);
 
       expect(mockRateLimiter.errorCode, 200);
       expect(mockRateLimiter.retryAfterHeader, isNull);